feat: move to ui subfolder
