+++ lib/ui/settings/about/about_view.dart
